+++ src/omero_browser_qt/__init__.py
@@ -54,4 +54,4 @@ __all__ = [
     "raw_pixels_store",
 ]
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"